fix: adapter_types underlying int32_t (wire protocol), fix cast helpers
- Change enum class adapter_types underlying type: int8_t → int32_t so
  the dataType field in mesh_message is a 4-byte sint32, matching the
  server's wire protocol expectations.
- Update static_assert: sizeof(mesh_message) 40 → 43 bytes (int32_t is
  4B vs int8_t 1B for the dataType field).
- Fix adapterTypeFromEEPROM/adapterTypeToEEPROM helpers in
  AdapterFactory.cpp to cast via int32_t instead of int8_t.
- Replace raw int8_t cast in Adapter.cpp CONFIG_SET handler with the
  named adapterTypeFromEEPROM() helper.
- Replace raw static_cast<uint8_t>(adapter_types::SERIAL_ADAPTER) in
  Serial_Adapter::sendHealthReport() with adapterTypeToEEPROM() helper.
- Update all Logger String() cast sites from int8_t → int32_t.

+++ main/src/Adapter/Adapter.h
@@ -14,7 +14,7 @@ namespace planetopia {
 namespace adapter {
 
 // Enum for identifying adapter types
-enum class adapter_types : int8_t {
+enum class adapter_types : int32_t {
   UNKNOWN_ADAPTER = -1,
   PIR_ADAPTER = 0,
   WIFI_ADAPTER = 1,
